requirements.txt
Django==5.2.12
djangorestframework
django-allauth